feat(home/content): implement home and content management
- Add `home` module for version management, settings, and game launch.
- Add `content-download` module for CurseForge and LIP integration.
- Implement module registration system for frontend router and nav.
- Add version icon editing and content management UI.
- Update UI tokens to copper branding and integrate new modules.

+++ src-tauri/src/modules/home/meta.rs
@@ -203,7 +203,7 @@ pub fn now_rfc3339() -> String {
         .map(|d| d.as_secs())
         .unwrap_or(0);
     let days = secs.div_euclid(86_400) as i64;
-    let rem = secs.rem_euclid(86_400) as u64;
+    let rem = secs.rem_euclid(86_400);
     let (hour, min, sec) = (rem / 3600, (rem % 3600) / 60, rem % 60);
     // civil_from_days: 天数 → 公历日期。
     let z = days + 719_468;
